+++ CHANGELOG.md
@@ -208,3 +208,4 @@
     - Bug-088726 - Add to cart button is appearing in case of RFQ only
     - Bug-088716 - The seller's name is visible in the RFQ form seller's suggestion listing even if the seller disables the RFQ at the shop level
     - Bug-088767 - RFQ >> COD is available on the front end even if it is disabled from the product setup form
+    - Bug-088768 - The volume discount and buy-together products should not appear on the product details page when the Hide Price setting is enabled